chapter-2: moves exception handlers in a dedicated class annotated with ControllerAdvice in order to extend their scope
